LayoutTests/platform/mac/accessibility/progressbar-expected.txt

X X  
This test makes sure that progress element can be accessed by Accessibility FW.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS obj.intValue is 7
PASS obj.intValue is 0
FAIL obj.intValue should be 7. Was 0.
FAIL obj.role should be AXRole: AXProgressIndicator. Was AXRole: AXGroup.
PASS obj.intValue is 0
FAIL obj.role should be AXRole: AXProgressIndicator. Was AXRole: AXGroup.
PASS successfullyParsed is true

TEST COMPLETE
